keep progress

+++ sample/crypto/gen.sh
@@ -6,12 +6,12 @@ PSW='&wA*+<_Afh2*4#Z'
 BIN=$(dirname "$([[ $0 == /* ]] && echo "$0" || echo "$PWD/${0#./}")")
 export PATH="${PATH}:/opt/mqm/bin"
 
-rm -rf "$BIN"/{client,server}
-mkdir "$BIN"/{client,server}
+rm -rf "${BIN:?}"/{client,server}
+mkdir "${BIN:?}"/{client,server}
 
-CACERT="${BIN}/server/cacert.crt"
-CAKEY="${BIN}/server/ca.key"
-CLIENT_DB="${BIN}/client/keys.kdb"
+CACERT="${BIN:?}/server/cacert.crt"
+CAKEY="${BIN:?}/server/ca.key"
+CLIENT_DB="${BIN:?}/client/keys.kdb"
 
 openssl req -newkey rsa:2048 -nodes -keyout "$CAKEY" -x509 -days 365 -out "$CACERT" \
   -subj "/C=RU/ST=MO/L=Moscow/O=VTB/OU=Finance/CN='mqpro-sample'/emailAddress=email@vtb.ru"
